+++ src/main/java/com/hiczp/bilibili/api/live/socket/entity/SysGiftEntity.java
@@ -26,9 +26,9 @@ public class SysGiftEntity {
     @SerializedName("url")
     private String url;
     @SerializedName("roomid")
-    private int roomId;
+    private long roomId;
     @SerializedName("real_roomid")
-    private int realRoomId;
+    private long realRoomId;
     @SerializedName("giftId")
     private int giftId;
     @SerializedName("msgTips")
@@ -74,19 +74,19 @@ public class SysGiftEntity {
         this.url = url;
     }
 
-    public int getRoomId() {
+    public long getRoomId() {
         return roomId;
     }
 
-    public void setRoomId(int roomId) {
+    public void setRoomId(long roomId) {
         this.roomId = roomId;
     }
 
-    public int getRealRoomId() {
+    public long getRealRoomId() {
         return realRoomId;
     }
 
-    public void setRealRoomId(int realRoomId) {
+    public void setRealRoomId(long realRoomId) {
         this.realRoomId = realRoomId;
     }
